writing-practice/requirements.txt
streamlit==1.31.0
openai==1.71.0
python-dotenv==1.0.1
instructor==1.0.0
requests>=2.32.3
PyYAML 